second commit

+++ CS321_W5D2_BlogAPI.Core/Services/PostService.cs
@@ -23,7 +23,19 @@ namespace CS321_W5D2_BlogAPI.Core.Services
             //     Use the _userService to get the current users id.
             //     You may have to retrieve the blog in order to check user id
             // TODO: assign the current date to DatePublished
-            return _postRepository.Add(newPost);
+            var currentUser = _userService.CurrentUserId;
+            var currentBlog = _blogRepository.Get(newPost.BlogId);
+            if (currentUser == currentBlog.UserId)
+            {
+                newPost.DatePublished = DateTime.Now;
+                return _postRepository.Add(newPost);
+            }
+            else
+            {
+                throw new ApplicationException("This is not your blog!");
+            }
+
+
         }
 
         public Post Get(int id)
@@ -44,14 +56,30 @@ namespace CS321_W5D2_BlogAPI.Core.Services
         public void Remove(int id)
         {
             var post = this.Get(id);
-            // TODO: prevent user from deleting from a blog that isn't theirs
-            _postRepository.Remove(id);
+            
+            if (_userService.CurrentUserId == post.Blog.UserId)
+            {
+                _postRepository.Remove(id);
+                return;
+            }
+            else
+            {
+                throw new ApplicationException("This is not your blog!");
+            }
         }
 
         public Post Update(Post updatedPost)
         {
-            // TODO: prevent user from updating a blog that isn't theirs
-            return _postRepository.Update(updatedPost);
+            var currentUser = _userService.CurrentUserId;
+            var blog = _blogRepository.Get(updatedPost.Id);
+            if (currentUser == blog.UserId)
+            {
+                return _postRepository.Update(updatedPost);
+            }
+            else
+            {
+                throw new ApplicationException("This is not your blog!");
+            }
         }
 
     }
